--- a/phalcon/events/manager.zep
+++ b/phalcon/events/manager.zep
@@ -20,7 +20,7 @@
 namespace Phalcon\Events;
 
 use Phalcon\Events\Event;
-use SplPriorityQueue as PriorityQueue;
+use SplPriorityQueue;
 
 /**
  * Phalcon\Events\Manager
@@ -61,10 +61,10 @@ class Manager implements ManagerInterface
 			if this->_enablePriorities {
 
 				// Create a SplPriorityQueue to store the events with priorities
-				let priorityQueue = new PriorityQueue();
+				let priorityQueue = new SplPriorityQueue();
 
 				// Extract only the Data // Set extraction flags
-				priorityQueue->setExtractFlags(PriorityQueue::EXTR_DATA);
+				priorityQueue->setExtractFlags(SplPriorityQueue::EXTR_DATA);
 
 				// Append the events to the queue
 				let this->_events[eventType] = priorityQueue;
@@ -104,10 +104,10 @@ class Manager implements ManagerInterface
 			if typeof priorityQueue == "object" {
 
 				// SplPriorityQueue hasn't method for element deletion, so we need to rebuild queue
-				let newPriorityQueue = new PriorityQueue();
-				newPriorityQueue->setExtractFlags(\SplPriorityQueue::EXTR_DATA);
+				let newPriorityQueue = new SplPriorityQueue();
+				newPriorityQueue->setExtractFlags(SplPriorityQueue::EXTR_DATA);
 
-				priorityQueue->setExtractFlags(PriorityQueue::EXTR_BOTH);
+				priorityQueue->setExtractFlags(SplPriorityQueue::EXTR_BOTH);
 				priorityQueue->top();
 
 				while priorityQueue->valid() {
@@ -201,7 +201,7 @@ class Manager implements ManagerInterface
 
 		if typeof queue != "array" {
 			if typeof queue == "object" {
-				if !(queue instanceof \SplPriorityQueue) {
+				if !(queue instanceof SplPriorityQueue) {
 					throw new Exception(
 						sprintf(
 							"Unexpected value type: expected object of type SplPriorityQueue, %s given",