… CHANGELOG-3.2.md
… tests/unit/Mvc/Dispatcher/DispatcherAfterExecuteRouteMethodTest.php
… tests/unit/Mvc/Dispatcher/DispatcherAfterExecuteRouteTest.php
… tests/unit/Mvc/Dispatcher/DispatcherAfterInitializeTest.php
… tests/unit/Mvc/Dispatcher/DispatcherBeforeDispatchLoopTest.php
… tests/unit/Mvc/Dispatcher/DispatcherBeforeDispatchTest.php
… tests/unit/Mvc/Dispatcher/DispatcherBeforeExecuteRouteMethodTest.php
… tests/unit/Mvc/Dispatcher/DispatcherBeforeExecuteRouteTest.php
… tests/unit/Mvc/Dispatcher/DispatcherInitalizeMethodTest.php
… tests/unit/Mvc/Dispatcher/Helper/BaseDispatcher.php
… tests/unit/Mvc/Dispatcher/Helper/DispatcherListener.php
… tests/unit/Mvc/Dispatcher/Helper/DispatcherTestAfterExecuteRouteForwardController.php
… tests/unit/Mvc/Dispatcher/Helper/DispatcherTestAfterExecuteRouteReturnFalseController.php
… tests/unit/Mvc/Dispatcher/Helper/DispatcherTestBeforeExecuteRouteExceptionController.php
… tests/unit/Mvc/Dispatcher/Helper/DispatcherTestBeforeExecuteRouteForwardController.php
… tests/unit/Mvc/Dispatcher/Helper/DispatcherTestBeforeExecuteRouteReturnFalseController.php
… tests/unit/Mvc/Dispatcher/Helper/DispatcherTestDefaultController.php
… tests/unit/Mvc/Dispatcher/Helper/DispatcherTestDefaultNoNamespaceController.php
… tests/unit/Mvc/Dispatcher/Helper/DispatcherTestDefaultSimpleController.php
… tests/unit/Mvc/Dispatcher/Helper/DispatcherTestDefaultTwoController.php
… tests/unit/Mvc/Dispatcher/Helper/DispatcherTestInitializeExceptionController.php
… tests/unit/Mvc/Dispatcher/Helper/DispatcherTestInitializeForwardController.php
… tests/unit/Mvc/Dispatcher/Helper/DispatcherTestInitializeReturnFalseController.php
… tests/unit/Mvc/DispatcherTest.php
… unit-tests/DispatcherMvcEventsTest.php