This is a an update of previously closed PR #12209.  It resolves all dispatch issues, adds additional notes in the code for future deprecations in 4.0. Various performance improvements in the resulting dispatch process as result of variable caching and dropping the global try/catch which previously hooked into the RVAL for every statement previously.

Fixes current issue: #12931

Related to (and ensures fixes for) previous issues:
#12154
#11819

cc @andresgutierrez
cc @temuri416 
cc @sergeyklay 
cc @niden 

```
Unit Tests (69) -------------------------------------------------------------------------------
✔ DispatcherAfterDispatchLoopTest: After dispatch loop forward (0.03s)
✔ DispatcherAfterDispatchLoopTest: After dispatch loop return false (0.00s)
✔ DispatcherAfterDispatchLoopTest: After dispatch loop with before exception returning false (0.00s)
✔ DispatcherAfterDispatchLoopTest: After dispatch loop with before exception bubble (0.00s)
✔ DispatcherAfterDispatchLoopTest: After dispatch loop with before exception forward once (0.00s)
✔ DispatcherAfterDispatchTest: After dispatch forward once (0.00s)
✔ DispatcherAfterDispatchTest: After dispatch return false (0.00s)
✔ DispatcherAfterDispatchTest: After dispatch with before exception returning false (0.00s)
✔ DispatcherAfterDispatchTest: After dispatch with before exception bubble (0.00s)
✔ DispatcherAfterDispatchTest: After dispatch with before exception forward once (0.00s)
✔ DispatcherAfterExecuteRouteMethodTest: After execute route forward once (0.00s)
✔ DispatcherAfterExecuteRouteMethodTest: After execute route return false (0.00s)
✔ DispatcherAfterExecuteRouteMethodTest: After execute route with before exception returning false (0.00s)
✔ DispatcherAfterExecuteRouteMethodTest: After execute route with before exception bubble (0.00s)
✔ DispatcherAfterExecuteRouteMethodTest: After execute route with before exception forward once (0.00s)
✔ DispatcherAfterExecuteRouteTest: After execute route forward once (0.00s)
✔ DispatcherAfterExecuteRouteTest: After execute route return false (0.00s)
✔ DispatcherAfterExecuteRouteTest: After execute route with before exception returning false (0.00s)
✔ DispatcherAfterExecuteRouteTest: After execute route with before exception bubble (0.00s)
✔ DispatcherAfterExecuteRouteTest: After execute route with before exception forward once (0.00s)
✔ DispatcherAfterInitializeTest: After initialize forward once (0.00s)
✔ DispatcherAfterInitializeTest: After initialize return false (0.00s)
✔ DispatcherAfterInitializeTest: After initialize with before exception returning false (0.00s)
✔ DispatcherAfterInitializeTest: After initialize with before exception bubble (0.00s)
✔ DispatcherAfterInitializeTest: After initialize with before exception forward once (0.00s)
✔ DispatcherBeforeDispatchLoopTest: Before dispatch loop forward (0.00s)
✔ DispatcherBeforeDispatchLoopTest: Before dispatch loop return false (0.00s)
✔ DispatcherBeforeDispatchLoopTest: Before dispatch loop with before exception returning false (0.00s)
✔ DispatcherBeforeDispatchLoopTest: Before dispatch loop with before exception bubble (0.00s)
✔ DispatcherBeforeDispatchLoopTest: Before dispatch loop with before exception forward (0.00s)
✔ DispatcherBeforeDispatchTest: Before dispatch forward once (0.00s)
✔ DispatcherBeforeDispatchTest: Before dispatch return false (0.00s)
✔ DispatcherBeforeDispatchTest: Before dispatch with before exception returning false (0.00s)
✔ DispatcherBeforeDispatchTest: Before dispatch with before exception bubble (0.00s)
✔ DispatcherBeforeDispatchTest: Before dispatch with before exception forward once (0.00s)
✔ DispatcherBeforeExecuteRouteMethodTest: Before execute route forward once (0.00s)
✔ DispatcherBeforeExecuteRouteMethodTest: Before execute route return false (0.00s)
✔ DispatcherBeforeExecuteRouteMethodTest: Before execute route with before exception returning false (0.00s)
✔ DispatcherBeforeExecuteRouteMethodTest: Before execute route with before exception bubble (0.00s)
✔ DispatcherBeforeExecuteRouteMethodTest: Before execute route with before exception forward (0.00s)
✔ DispatcherBeforeExecuteRouteTest: Before execute route forward once (0.00s)
✔ DispatcherBeforeExecuteRouteTest: Before execute route return false (0.00s)
✔ DispatcherBeforeExecuteRouteTest: Before execute route with before exception returning false (0.00s)
✔ DispatcherBeforeExecuteRouteTest: Before execute route with before exception bubble (0.00s)
✔ DispatcherBeforeExecuteRouteTest: Before execute route with before exception forward once (0.00s)
✔ DispatcherInitalizeMethodTest: Initialize forward (0.00s)
✔ DispatcherInitalizeMethodTest: Initialize return false (0.00s)
✔ DispatcherInitalizeMethodTest: Initialize with before exception returning false (0.00s)
✔ DispatcherInitalizeMethodTest: Initialize with before exception bubble (0.00s)
✔ DispatcherInitalizeMethodTest: Initialize with before exception forward (0.00s)
✔ DispatcherTest: Default dispatch loop events (0.00s)
✔ DispatcherTest: Default dispatch loop events with no handlers (0.00s)
✔ DispatcherTest: Controller action local forward (0.00s)
✔ DispatcherTest: Controller action external forward (0.00s)
✔ DispatcherTest: Controller action return value string (0.00s)
✔ DispatcherTest: Controller action return value int (0.00s)
✔ DispatcherTest: Params and return value (0.00s)
✔ DispatcherTest: Cyclical routing (0.00s)
✔ DispatcherTest: Handler not found (0.05s)
✔ DispatcherTest: Handler invalid (0.00s)
✔ DispatcherTest: Handler action not found (0.00s)
✔ DispatcherTest: Last handler (0.00s)
✔ DispatcherTest: Last handler forward (0.00s)
✔ DispatcherTest: No namespaces (0.00s)
✔ DispatcherTest: Mixing namespace forward (0.00s)
✔ DispatcherTest: Default resolve (0.02s)
✔ DispatcherTest: Manual call action (0.00s)
✔ DispatcherTest: Last handler exception forward (0.00s)
✔ DispatcherTest: Exception in before exception (0.00s)
-------------------------------------------------------------------------------------------

Time: 2.23 seconds, Memory: 12.00MB

OK (69 tests, 173 assertions)
```