@sergeyklay 2017-10-08
Good job. Could you deal with failed test
https://github.com/phalcon/cphalcon/blob/master/unit-tests/DispatcherMvcEventsTest.php#L206-L209