@virgofx 2017-10-08
For clarity --- the test expected value is marked in green and the new behavior is in red (Wording backwards)...

The reason that failed was because of a nuance in the old test. Specifically, it did not contain an **initialize()** in the controller in the old case and thus it should never fire the `afterInitialize`.  I can see this both ways ... the change was made to interpret the following logic:

* Event should be thrown right before/after the corresponding action is about to happen. E.g. in the case of the view, if it's disabled the view events are not thrown.  

My original thought is that the existing behavior is incorrect; however, I would be compelled to add a new test case to handle this. E.g. If the use case was to somehow globally instantiate controllers from this method via some common functionality.

In the case of the dispatcher, it depends how you read the **afterInitialize** event. Do you:

A) interpret *initialize* as being specific to initializing via the `initialize()` method ... or
B) refer to it as being general such that the event should be fired in it's relative location even without a direct call into the `initialize()` - e.g. The controller is now initialized.


Side note --- there is a separate note in the code. The whole [ initialize/afterInitialize ] block is in the wrong location. The beforeExecuteRoute should never be executed before the initialize; however, this was a bug in the original code/documentation. I left it in place currently to not break B/C; however, it would be worth moving this section such that the `initialize` handles certain constructor/core controller logic that is outside of the main action portion.