@sergeyklay 2017-10-08
Yes, the test expected value is marked in green and the new behavior is in red . I just copied phpunit output from the terminal. Well, could you just remove wrong test?