@virgofx 2017-10-08
@sergeyklay I updated original comment with more info and questions. It is removed as noted now (hence passing); however, do you have thoughts on [A] vs. [B] above? I'm open to reverting back to the [B] model as I could see there be a global use case for performing actions from an eventhook globally in controllers regardless of the presence of `initialize()`