@sergeyklay 2017-10-08
Actually I prefer [B] but in general your PR look really great. I will keep this open for a day or two to allow other contributors a chance to speak up, and to take a fresh look at it later. But at the moment it seems good to me.

cc: @phalcon/core-team