@niden 2017-10-09
This looks great. I would also prefer [B]. 

Great work buddy, thank you so much for the contribution!