@virgofx 2017-10-09
Modified to include option [B] above. This means:
* Patch is 100% backwards compatible (with proper bugfixes only)
* The legacy `unit-tests` that failed now passes (even though it doesn't exist anymore). Replaced by the new suite and specifically one of the core dispatcher tests: https://github.com/virgofx/cphalcon/blob/da5f3e0736263ed8eba7505c11c644286f90aee3/tests/unit/Mvc/Dispatcher/DispatcherTest.php#L93-L104
* A few documentation updates forthcoming to include some things that are not able to fix in this version.  (For example returning `false` does not work when there are more event listeners. Currently, the event must be stopped. Need to outline how to do this in the documentation for 3rd party applications/libraries that may publish plugins that hook into events and need to stop dispatch actions. The only sure fire way would be to return false AND stop the event). This could be potentially reworked for 4.0.