@niden 2017-10-09
For 3.x we should keep things the way they are. If I need events to stop (subsequent listeners do not fire) we will need to do `$event->stop()` and `return false`. This has to be documented for 3.x

For 4.x this behavior needs to change. Simply `return false` should suffice to stop the execution of subsequent listeners to the same event.

The above will be OK behavior so long as we ensure that we highlight it in the documentation.