@sergeyklay 2017-10-10
We're plan to release 3.2.3 in next days and start 3.3.0 branch. You can cherry-pick these changes to 4.0.x branch or leave this job for me. I'll do this later (but can't say when exactly).