@virgofx 2017-10-10
Doesn't matter to me ... I can just help with the 4.0 changes related to this feature. Is the 4.0 branch working/building and is it PHP7.x only ?  What would you prefer I do ?