@sergeyklay 2017-10-10
Yes, we'll drop support of the Zend Engine 2.x from Phalcon 4.x.