@temuri416 2017-10-26
@virgofx @sergeyklay 

It looks like this patch needs a small review.

It broke an existing behaviour: after forwarding to an action within the same controller - its `initialize()` method does not get called after forward.

However, when forwarding to a different controller, controller's `initialize()` does get invoked again.

```php
<?php

use Phalcon\Mvc\Controller;

class IndexController extends Controller
{
    public function initialize()
    {
        $this->view->location = $this->di->get('dispatcher')->getHandlerClass();
        $this->view->action = $this->di->get('dispatcher')->getActionName();
    }

    public function indexAction()
    {
        /**
         * Before #13112 this forward will result in "Your location: IndexController/welcome".
         *
         * After #13112 this forward will result in "Your location: IndexController/index".
         *
         * @var IndexController
         */
        $this->di['dispatcher']->forward([
            'controller' => 'index',
            'action' => 'welcome',
        ]);
    }

    public function welcomeAction()
    {
    }
}
```

This line seems to be the key: https://github.com/phalcon/cphalcon/blob/v3.2.3/phalcon/dispatcher.zep#L520.

Current controller is no longer considered a "fresh" instance.

Code to reproduce is attached.
[issue-13112.zip](https://github.com/phalcon/cphalcon/files/1418972/issue-13112.zip)

Thanks!