@virgofx 2017-10-26
@temuri416 That's the correct behavior. The initialize() is only called ONCE ever per controller. It's more of a static onConstruct() .. Right now it maps to how the models were done with initialize as well in the sense that 

It's a TERRIBLE implementation; however. I made tons of notes in the code. It doesn't make sense in it's current location being called AFTER beforeExecuteRoute().

Also .. this is how things were in the past (although there were other bugs that may have blocked the intended behavior and thus it may appear different)

The next version really needs to fix the wording. "Initialize" is confusing:

1) Next version **WILL** move that logic **BEFORE** the `beforeExecuteRoute()` handlers.

I believe the `initialize()` is correct in that it should only be executed once per Dispatcher/Di setup. It's a one-time initialization --- stuff like doing ACL for controllers is where this should happen.   If you need something to happen EVERY time in some sort of cyclical cycle -- that should happen in the `beforeExecuteRoute`