@sergeyklay 2017-10-26
@virgofx These changes already in 3.3.x branch. I'll deal with 4.0.0 a bit later